Fixed: GetPlayerCount & es_ForceValue
Fixed a bug with the GetPlayerCount command which was always returning the value null, similar to the previous fix with GetLivingPlayerCount.

Fixed: A problem with the command es_forcevalue not working. (Thanks to Xiazee for fixing this bug. now we are once again able to force values such as sv_cheat's this is huge!)