Fixed: Command ForceValue
Fixed the forcevalue command, it should now work and allow players to bypass sv_cheats features such as for example noclip. It does however announce that sv_cheats was changed to 1 and then back to 0 upon these commands execution. But so did it back in CS:S with Mattie's Eventscripts.
Thanks to XIazee for coming up with a solution and providing this fix.